% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brasilia_df}
\alias{brasilia_df}
\title{Monthly relative humidity in Brasília (data frame)}
\format{
A data frame with 2 columns:
\describe{
  \item{time}{Month of observation (\code{yearmon}).}
  \item{y}{Relative humidity as a proportion (numeric, in (0, 1)).}
}
}
\source{
NASA Prediction of Worldwide Energy Resources (POWER) project.
  \url{https://power.larc.nasa.gov/}
}
\usage{
brasilia_df
}
\description{
The same series as \code{\link{brasilia_ts}}, stored as a data frame
with a \code{yearmon} time column.
}
\examples{
data(brasilia_df)
head(brasilia_df)
}
\seealso{
\code{\link{brasilia_ts}} for the same data as a \code{ts} object.
}
\keyword{datasets}
