% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brasilia_ts}
\alias{brasilia_ts}
\title{Monthly relative humidity in Brasília}
\format{
A \code{ts} object with monthly observations, frequency 12,
  starting in January 1999.
}
\source{
NASA Prediction of Worldwide Energy Resources (POWER) project.
  \url{https://power.larc.nasa.gov/}
}
\usage{
brasilia_ts
}
\description{
Monthly relative humidity in Brasília, Brazil, expressed as a proportion.
The data were obtained from the NASA POWER project and cover
January 1999 to June 2024.
}
\examples{
data(brasilia_ts)
plot(brasilia_ts, ylab = "Relative humidity (proportion)", xlab = "Year")
}
\references{
Cribari-Neto, F., Costa, E., & Fonseca, R. V. (2025). Numerical stability
enhancements in beta autoregressive moving average model estimation.
\emph{Brazilian Journal of Probability and Statistics}, 39(4), 410--437.
\doi{10.1214/25-BJPS645}
}
\seealso{
\code{\link{brasilia_df}} for the same data as a data frame.
}
\keyword{datasets}
