% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.barma.R
\name{coef.barma}
\alias{coef.barma}
\title{Extract Coefficients from a barma Model}
\usage{
\method{coef}{barma}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class `"barma"`.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A named numeric vector of all estimated coefficients 
(e.g., alpha, varphi, theta, phi).
}
\description{
S3 method for extracting the vector of estimated coefficients from a fitted 
model object of class `"barma"`.
}
