% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.barma.R
\name{forecast.barma}
\alias{forecast.barma}
\title{Forecast a barma Model}
\usage{
\method{forecast}{barma}(object, h = 6, ...)
}
\arguments{
\item{object}{A fitted model object of class `"barma"`.}

\item{h}{The number of steps to forecast ahead (forecast horizon).
Default is 6.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A `ts` object containing the point forecasts for $h$ steps ahead.
}
\description{
S3 method for producing forecasts from a fitted `"barma"` model object.
}
\details{
This function computes dynamic, multi-step-ahead point forecasts, where 
future predictor-scale errors are assumed to be zero.

The autoregressive part of the forecast uses the conditional expectation
\eqn{E[g(y_t) | F_{t-1}] = \eta_t}, so future \eqn{g(y)} values are 
replaced by their forecasted \eqn{\eta_t} values.
}
