% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.barma.R
\name{print.summary.barma}
\alias{print.summary.barma}
\title{Print Method for a barma Summary}
\usage{
\method{print}{summary.barma}(x, ...)
}
\arguments{
\item{x}{A fitted model summary object of class `"summary.barma"`.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisibly returns the original object `x`.
}
\description{
S3 method for printing the detailed summary of a `"barma"` model object.
}
\details{
This function formats and displays the summary list created by
`summary.barma()`, including the call, coefficient table, and
information criteria.
}
