% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.barma.R
\name{summary.barma}
\alias{summary.barma}
\title{Summarize a barma Model Fit}
\usage{
\method{summary}{barma}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class `"barma"`.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A list object of class `"summary.barma"`...
}
\description{
(Full documentation block...)
}
