% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fim_barma.R
\name{fim_barma}
\alias{fim_barma}
\title{FIM and Fitted Values for a BARMA Model (Internal)}
\usage{
fim_barma(y, ar, ma, alpha, varphi, theta, phi, link)
}
\arguments{
\item{y}{A numeric vector representing the time series data, in (0, 1).}

\item{ar}{A numeric vector specifying the autoregressive (AR) lags.}

\item{ma}{A numeric vector specifying the moving average (MA) lags.}

\item{alpha}{The intercept term.}

\item{varphi}{A numeric vector of autoregressive (AR) parameters.}

\item{theta}{A numeric vector of moving average (MA) parameters.}

\item{phi}{The precision parameter of the BARMA model.}

\item{link}{The link function to be used.}
}
\value{
A list with the following elements:
  - 'fisher_info_mat': The Fisher information matrix.
  - 'fitted_ts': The fitted values (conditional mean) as a 'ts' object.
  - 'muhat_effective': The effective vector of fitted conditional means.
  - 'etahat_full': The full vector (NA-padded) of the linear predictor.
  - 'errorhat_full': The full vector of errors on the predictor scale.
}
\description{
This function computes the Fisher information matrix of the
  Beta Autoregressive Moving Average (BARMA) model. It also efficiently
  returns auxiliary values like fitted values and residuals required
  by the main 'barma' function.
}
\keyword{internal}
