% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.barma.R
\name{forecast.barma}
\alias{forecast.barma}
\title{Forecast a barma Model}
\usage{
\method{forecast}{barma}(object, h = 6, xreg = NULL, ...)
}
\arguments{
\item{object}{A fitted model object of class `barma`. Must contain
`object$xreg` if regressors were used.}

\item{h}{The number of steps to forecast ahead (forecast horizon).
Default is 6.}

\item{xreg}{A matrix of future regressor values for the forecast horizon.
Should have h rows and the same number of columns as xreg used in model
fitting.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A `ts` object containing the point forecasts for h steps ahead.
}
\description{
S3 method for producing forecasts from a fitted `barma` model object.
}
\details{
This function computes dynamic, multi-step-ahead point forecasts.
It implements a "Regression with ARMA errors" logic, where the AR components
are applied to the deviations from the regression line:
\eqn{AR(g(y_{t-k}) - x_{t-k}^\top \beta)}.
}
